feature/R15/t02/ - Conflitos gerados pela merge Esse commit foi feito para arrumar erros gerados pelo merge com a branch principal com o trabalho que estava sendo feito. - Realocação dos arquivos - Adequação para padronização de codigo para a inserção dos dados do produto. - Mudanças nas classes de field para melhor escrita de codigo.
